LayoutTests/fast/forms/datetimelocal/datetimelocal-editable-components/picker-keyboard-accessibility-expected.txt

Tests that focus is properly managed when alternating between activating the picker with the keyboard and with a mouse click.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS showingPicker is true
PASS input.value is "2020-09-17T10:45"
PASS input.value is "2020-08-17T10:45"
PASS input.value is "2020-08-16T10:45"
PASS showingPicker is false
PASS input.value is "2019-08-16T10:45"
PASS successfullyParsed is true

TEST COMPLETE
